+++ server/prisma/migrations/20260701155432_two_factor_email_otp/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "login_token_type" ADD VALUE 'TWO_FACTOR_OTP';